# backend/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /users/{username}:
    get:
      description: Retrieve user profile by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by username
      tags:
      - users
  /users/me:
    get:
      description: Retrieve your own user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get your profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update your display name and avatar
      parameters:
      - description: Profile updates
        in: body
        name: request
        required: true
        schema:
          properties:
            avatar_url:
              type: string
            display_name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - users
  /users/search:
    get:
      description: Search for users by username
      parameters:
      - description: Search query (min 2 chars)
        in: query
        name: q
        required: true
        type: string
      - description: Result limit (default 20, max 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              users:
                items: {}
                type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search users
      tags:
      - users
swagger: "2.0"